requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
asyncio-atexit==1.0.1
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
emoji==2.14.1
fastapi==0.116.1
grapheme==0.6.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
mss==10.1.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
ua-generator==2.0.11
uvicorn==0.35.0
websockets==15.0.1
wrapt==1.17.3
zendriver==0.13.1
